// moonpad/src/trace-types.ts
/*
 * Copyright 2025 International Digital Economy Academy
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

type TraceResult = {
  name: string;
  value: string;
  line: number;
  start_column: number;
  end_column: number;
  filepath?: string;
  hit: number;
};

type TraceRunOutput =
  | {
      kind: "trace-delta";
      entries: TraceResult[];
    }
  | {
      kind: "stdout-batch";
      lines: string[];
    };

export type { TraceResult, TraceRunOutput };